// ui/src/api/generated/model/plantCreateRequest.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Whirlpool Inspection Tool API
 * APIs for Whirlpool Inspection Tool API developed by Scopt Analytics for Whirlpool Inspection Tool API platform.
 * OpenAPI spec version: 1.0.3
 */

export interface PlantCreateRequest {
  /**
     * @minLength 2
     * @maxLength 64
     */
  plant_code: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  lat: number;
  lng: number;
  /**
     * @minLength 2
     * @maxLength 128
     */
  address: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  city: string;
  /** @minimum 0 */
  postal_code: number;
}