// ui/src/api/generated/model/plantResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Whirlpool PDI Tool API
 * Whirlpool PDI Tool API APIs developed by Scopt Analytics.
 * OpenAPI spec version: 1.6.0
 */
import type { FacilityStatsResponse } from './facilityStatsResponse';

export interface PlantResponse {
  id: number;
  uuid: string;
  plant_code: string;
  name: string;
  lat: number | null;
  lng: number | null;
  address: string;
  city: string;
  postal_code: string;
  is_active: boolean;
  created_at: string;
  updated_at: string;
  stats: FacilityStatsResponse;
}